pyproject.toml
[tool.poetry]
name = "fuel-uk"
version = "0.1.0"
description = ""
authors = ["Dominik Golebiewski <67642192+DominikGolebiewski@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
requests = "^2.32.3"
python-slugify = "^8.0.4"
duckdb = "^1.0.0"
streamlit = "^1.36.0"
folium = "^0.17.0"
streamlit-folium = "^0.21.1"
boto3 = "^1.34.145"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
